Improves registration process for group classes
Enhances the registration workflow to accommodate both private and group classes.

This change introduces the concept of class-based registrations, which previously only supported private lessons.

Updates the registration form to allow selection of class times for group lessons.

The registration id generation accounts for group lessons which are classId based.