Refactors authentication middleware
Simplifies authentication by removing the complex middleware flow and replacing it with repository initialization only.
Authentication and authorization are now handled within the controller logic of specific API endpoints.

This change ensures that repositories are always available to the API, while authentication checks are performed explicitly where needed.
This approach provides more flexibility and control over authentication, especially for operator user retrieval, and reduces reliance on environment variables.

The previous authentication middleware had problems.  For example, the operator email would have to be present in the roles table, and the corresponding access codes in the admin, instructor or parent tables.  This has been simplified.